[#825] Extracted Drupal AJAX handling into 'DrupalAjaxTrait'.
