+++ src/Drupal/DrupalExtension/Context/MinkContext.php
@@ -6,17 +6,14 @@ namespace Drupal\DrupalExtension\Context;
 
 use Behat\Step\Given;
 use Behat\Step\When;
-use Behat\Hook\BeforeStep;
-use Behat\Hook\AfterStep;
 use Behat\Step\Then;
-use Behat\Behat\Hook\Scope\BeforeStepScope;
-use Behat\Behat\Hook\Scope\AfterStepScope;
 use Behat\Behat\Context\TranslatableContext;
 use Behat\Mink\Exception\ElementNotFoundException;
 use Behat\Mink\Exception\ExpectationException;
 use Behat\Mink\Exception\UnsupportedDriverActionException;
 use Behat\Mink\Selector\Xpath\Escaper;
 use Behat\MinkExtension\Context\MinkContext as MinkExtension;
+use Drupal\DrupalExtension\Context\Traits\DrupalAjaxTrait;
 use Drupal\DrupalExtension\ParametersTrait;
 use Drupal\DrupalExtension\RegionTrait;
 use Drupal\DrupalExtension\TagTrait;
@@ -26,6 +23,7 @@ use Drupal\DrupalExtension\TagTrait;
  */
 class MinkContext extends MinkExtension implements TranslatableContext, ParametersAwareInterface {
 
+  use DrupalAjaxTrait;
   use ParametersTrait;
   use RegionTrait;
   use TagTrait;
@@ -108,94 +106,6 @@ class MinkContext extends MinkExtension implements TranslatableContext, Paramete
   public function iEnterValueForField(string $value, string $field): void {
     // Use the Mink Extension step definition.
     $this->fillField($field, $value);
-  }
-
-  /**
-   * For javascript enabled scenarios, always wait for AJAX before clicking.
-   */
-  #[BeforeStep]
-  public function beforeJavascriptStep(BeforeStepScope $event): void {
-    /** @var \Behat\Behat\Hook\Scope\BeforeStepScope $event */
-    // Make sure the feature is registered in case this hook fires before
-    // ::registerFeature() which is also a @BeforeStep. Behat doesn't
-    // support ordering hooks.
-    $this->registerFeature($event);
-    if (!$this->hasTag('javascript')) {
-      return;
-    }
-    $text = $event->getStep()->getText();
-    if (preg_match('/\b(follow|press|click|submit|attach)\b/i', $text)) {
-      $this->iWaitForAjaxToFinish($event);
-    }
-  }
-
-  /**
-   * For javascript enabled scenarios, always wait for AJAX after clicking.
-   */
-  #[AfterStep]
-  public function afterJavascriptStep(AfterStepScope $event): void {
-    if (!$this->hasTag('javascript')) {
-      return;
-    }
-    $text = $event->getStep()->getText();
-    if (preg_match('/\b(follow|press|click|submit|attach)\b/i', $text)) {
-      $this->iWaitForAjaxToFinish($event);
-    }
-  }
-
-  /**
-   * Wait for AJAX to finish.
-   *
-   * @see \Drupal\FunctionalJavascriptTests\JSWebAssert::assertWaitOnAjaxRequest()
-   *
-   * @code
-   * Given I wait for AJAX to finish
-   * @endcode
-   */
-  #[Given('I wait for AJAX to finish')]
-  public function iWaitForAjaxToFinish(mixed $event = NULL): void {
-    if (!$this->getSession()->isStarted()) {
-      return;
-    }
-
-    $condition = <<<JS
-    (function() {
-      function isAjaxing(instance) {
-        return instance && instance.ajaxing === true;
-      }
-      var d7_not_ajaxing = true;
-      if (typeof Drupal !== 'undefined' && typeof Drupal.ajax !== 'undefined' && typeof Drupal.ajax.instances === 'undefined') {
-        for(var i in Drupal.ajax) { if (isAjaxing(Drupal.ajax[i])) { d7_not_ajaxing = false; } }
-      }
-      var d8_not_ajaxing = (typeof Drupal === 'undefined' || typeof Drupal.ajax === 'undefined' || typeof Drupal.ajax.instances === 'undefined' || !Drupal.ajax.instances.some(isAjaxing))
-      return (
-        // Assert no AJAX request is running (via jQuery or Drupal) and no
-        // animation is running.
-        (typeof jQuery === 'undefined' || jQuery.hasOwnProperty('active') === false || (jQuery.active <= 0 && jQuery(':animated').length === 0)) &&
-        d7_not_ajaxing && d8_not_ajaxing
-      );
-    }());
-JS;
-    $ajax_timeout = $this->getParameter('ajax_timeout');
-    $result = $this->getSession()->wait(1000 * $ajax_timeout, $condition);
-    if (!$result) {
-      if ($ajax_timeout === NULL) {
-        throw new \RuntimeException('No AJAX timeout has been defined. Please verify that "Drupal\DrupalExtension" is configured in behat.yml.');
-      }
-      if ($event) {
-        /** @var \Behat\Behat\Hook\Scope\BeforeStepScope $event */
-        $event_data = ' ' . json_encode([
-          'name' => $event->getName(),
-          'feature' => $event->getFeature()->getTitle(),
-          'step' => $event->getStep()->getText(),
-          'suite' => $event->getSuite()->getName(),
-        ]);
-      }
-      else {
-        $event_data = '';
-      }
-      throw new \RuntimeException('Unable to complete AJAX request.' . $event_data);
-    }
   }
 
   /**
